requirements.txt
django
djangorestframework
django-ninja
pymupdf
python-Levenshtein
nltk
transformers
torch
torchvision
torchaudio